+++ 2023/otel/README.md
@@ -7,3 +7,6 @@ https://opentelemetry.io/docs/instrumentation/python/getting-started/
 https://opentelemetry.io/docs/instrumentation/python/automatic/
 
 http://localhost:8000/roll?sides=2&rolls=6
+
+Looks like FastAPI doesn't have good auto instrumentation for exceptions yet:
+https://medium.com/humanmanaged/quick-start-to-integrating-opentelemetry-with-fastapi-part-1-2be4fec874bc
